legacy/src/preprocess/requirements.txt
pandas==1.5.3
cycler==0.12.1
fonttools==4.55.3
iniconfig==2.0.0
joblib==1.4.2
kiwisolver==1.3.2
numpy==1.22.4
nvidia-nccl-cu12==2.24.3
packaging==24.2
pillow==11.1.0
plotly==5.24.1
pluggy==1.5.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.0
scipy==1.13.1
six==1.17.0
tenacity==9.0.0
threadpoolctl==3.5.0
tzdata==2024.2
xgboost==2.1.3
matplotlib==3.5.3
dvc==2.6.0
pyparsing==2.4.7
fsspec==2022.5.0
PyYAML==6.0
